:sparkles: SC-4: add transformers for swagger UI customization
