+++ springdoc-customizer-common/src/main/java/io/funky/fangs/springdoc/customizer/annotation/ExampleMethod.java
