+++ springdoc-customizer-common/src/main/java/io/funky/fangs/springdoc/customizer/utility/ExampleUtilities.java
