+++ springdoc-customizer-common/src/main/resources/META-INF/services/javax.annotation.processing.Processor
