+++ springdoc-customizer-common/src/test/java/io/funky/fangs/springdoc/customizer/processor/ExamplesProcessorTest.java
