+++ springdoc-customizer-common/src/test/java/io/funky/fangs/springdoc/customizer/utility/ExampleUtilitiesTest.java
